Fix various mouse related issues
Fixed a bug where colored squares were clickable and draggable.

Created additional event listeners to disable coloring when the mouse leaves "container" div. This was causing an issue that kept the coloring toggle state active even if the mouse button was no longer held down outside of the sketch area.